move data source filter functions to data_source_meta
