+++ corehq/apps/userreports/app_manager/data_source_meta.py
@@ -1,4 +1,4 @@
-
+from django.utils.translation import ugettext_lazy as _
 
 DATA_SOURCE_TYPE_CASE = 'case'
 DATA_SOURCE_TYPE_FORM = 'form'
@@ -15,3 +15,27 @@ DATA_SOURCE_DOC_TYPE_MAPPING = {
 
 def get_data_source_doc_type(data_source_type):
     return DATA_SOURCE_DOC_TYPE_MAPPING[data_source_type]
+
+
+def make_case_data_source_filter(case_type):
+    return {
+        "type": "boolean_expression",
+        "operator": "eq",
+        "expression": {
+            "type": "property_name",
+            "property_name": "type"
+        },
+        "property_value": case_type,
+    }
+
+
+def make_form_data_source_filter(xmlns):
+    return {
+        "type": "boolean_expression",
+        "operator": "eq",
+        "expression": {
+            "type": "property_name",
+            "property_name": "xmlns"
+        },
+        "property_value": xmlns,
+    }
